mail_receiver/requirements.txt
aiosmtpd==1.4.4.post2
python-decouple==3.8
redis==5.0.1
